requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
requests==2.31.0
Pillow==10.0.1
PyPDF2==3.0.1
python-docx==0.8.11
xmltodict==0.13.0
openai==1.95.1
azure-identity==1.15.0
azure-mgmt-resource==23.0.1
azure-mgmt-authorization==4.0.0
zipfile36==0.1.3
python-dotenv==1.0.0
gunicorn==21.2.0
